Add config to disable plugin for more apps
When running fzf or other tools, it's nice if the plugin doesn't move
panes or tabs when using the ctrl+h/j/k/l shortcuts, because those tools
also use those shortcuts for navigation.

This change adds a configuration option to not switch panes or tabs when
in the listed disabled apps, by default vim, nvim and fzf. People can
add more apps if they need to.