Add enable and disable messages
Allow people to control if the plugin is active or not. This can be
helpful when the plugin needs to be temporarily disabled when an
application uses the ctrl/alt+hjkl keybindings and you wouldn't want the
plugin to switch pane or tab.

Updated the README with an example use case I ran into myself.